Rename storage to data package and move files around
