+++ libraries/core/src/main/java/com/thirdwavelist/coficiando/core/data/db/Database.kt
@@ -1,10 +1,10 @@
-package com.thirdwavelist.coficiando.core.storage.db
+package com.thirdwavelist.coficiando.core.data.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.thirdwavelist.coficiando.core.storage.db.cafe.CafeDao
-import com.thirdwavelist.coficiando.core.storage.db.cafe.CafeItem
+import com.thirdwavelist.coficiando.core.data.db.cafe.CafeDao
+import com.thirdwavelist.coficiando.core.data.db.cafe.CafeItem
 
 @Database(version = 1, entities = [CafeItem::class])
 @TypeConverters(CustomTypeConverters::class)
