+++ libraries/core/src/main/java/com/thirdwavelist/coficiando/core/data/db/cafe/BeanOriginType.kt
@@ -1,0 +1,6 @@
+package com.thirdwavelist.coficiando.core.data.db.cafe
+
+enum class BeanOriginType {
+    SINGLE,
+    BLEND
+}
