+++ libraries/core/src/main/java/com/thirdwavelist/coficiando/core/data/db/cafe/CafeItem.kt
@@ -1,4 +1,4 @@
-package com.thirdwavelist.coficiando.core.storage.db.cafe
+package com.thirdwavelist.coficiando.core.data.db.cafe
 
 import android.net.Uri
 import androidx.room.ColumnInfo
