+++ libraries/core/src/main/java/com/thirdwavelist/coficiando/core/data/db/cafe/GearInfoItem.kt
@@ -1,3 +1,3 @@
-package com.thirdwavelist.coficiando.core.storage.db.cafe
+package com.thirdwavelist.coficiando.core.data.db.cafe
 
 data class GearInfoItem(val espressoMachineName: String?, val grinderMachineName: String?)
