+++ libraries/core/src/main/java/com/thirdwavelist/coficiando/core/di/modules/StorageModule.kt
@@ -2,8 +2,8 @@ package com.thirdwavelist.coficiando.core.di.modules
 
 import android.content.Context
 import androidx.room.Room
-import com.thirdwavelist.coficiando.core.storage.db.Database
-import com.thirdwavelist.coficiando.core.storage.db.cafe.CafeDao
+import com.thirdwavelist.coficiando.core.data.db.Database
+import com.thirdwavelist.coficiando.core.data.db.cafe.CafeDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
