--- libraries/core/src/main/java/com/thirdwavelist/coficiando/core/storage/db/cafe/BeanOriginType.kt
@@ -1,6 +1,0 @@
-package com.thirdwavelist.coficiando.core.storage.db.cafe
-
-enum class BeanOriginType {
-    SINGLE,
-    BLEND
-}
